feat: introduces Kustomize support for Feature manifests
Enables use of a Feature API with Kustomize manifests.
They can be now used next to resources defined as YAML files or templates.
This makes Feature API able to handle component manifests in the same
fashion as infrastructure setup in use today.

Feature can have multiple KustomizeManifests defined through subsequent
calls of Kustomize() sub-builder.

There is a possibility to define global list of plugins, which will be applied for each of those.
This is used by the FeatureHandler to ensure both target namespace and component labels
(if applicable) will be added to Kustomize resources before applied to the cluster.

This is stacked PR for authorization capability.

Co-authored-by: Cameron Garrison <cgarriso@redhat.com>